review comments waitid (SQUASH)
waitpid is required by POSIX to be a cancellation point.  pclose is
specified as undefined if a cancellation occurs, so it would be
permitted for it to call a cancellable wait function; however, as a
quality of implementation matter, pclose must close the pipe fd before
it can wait (consider popen("yes","r")) and if the wait could be
interrupted the pipe FILE would be left in an intermediate state that
portable software cannot recover from, so the only useful behavior is
for pclose to NOT be a cancellation point.  We therefore support both at
a small cost in code size.

wait4 is historically not a cancellation point in musl; we retain that
since we need the non-cancellable version of __wait4 anyway.